move IndexType
