+++ core/src/index/knowhere/knowhere/index/IndexType.cpp
@@ -9,9 +9,10 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License
 
-#include "knowhere/index/vector_index/IndexType.h"
 #include <unordered_map>
+
 #include "knowhere/common/Exception.h"
+#include "knowhere/index/IndexType.h"
 
 namespace milvus {
 namespace knowhere {
