+++ core/src/index/knowhere/knowhere/index/vector_index/ConfAdapterMgr.h
@@ -15,8 +15,8 @@
 #include <memory>
 #include <unordered_map>
 
+#include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/ConfAdapter.h"
-#include "knowhere/index/vector_index/IndexType.h"
 
 namespace milvus {
 namespace knowhere {
