+++ core/src/index/knowhere/knowhere/index/vector_index/FaissBaseBinaryIndex.h
@@ -18,7 +18,7 @@
 
 #include "knowhere/common/BinarySet.h"
 #include "knowhere/common/Dataset.h"
-#include "knowhere/index/vector_index/IndexType.h"
+#include "knowhere/index/IndexType.h"
 
 namespace milvus {
 namespace knowhere {
