+++ core/src/index/knowhere/knowhere/index/vector_index/FaissBaseIndex.cpp
@@ -13,8 +13,8 @@
 #include <fiu-local.h>
 
 #include "knowhere/common/Exception.h"
+#include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/FaissBaseIndex.h"
-#include "knowhere/index/vector_index/IndexType.h"
 #include "knowhere/index/vector_index/helpers/FaissIO.h"
 
 namespace milvus {
