+++ core/src/index/knowhere/knowhere/index/vector_index/VecIndexFactory.h
@@ -13,7 +13,7 @@
 
 #include <memory>
 
-#include "knowhere/index/vector_index/IndexType.h"
+#include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/VecIndex.h"
 
 namespace milvus {
