+++ core/src/index/knowhere/knowhere/index/vector_index/gpu/IndexGPUIDMAP.cpp
@@ -19,8 +19,8 @@
 #include <fiu-local.h>
 
 #include "knowhere/common/Exception.h"
+#include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/IndexIDMAP.h"
-#include "knowhere/index/vector_index/IndexType.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
 #include "knowhere/index/vector_index/gpu/IndexGPUIDMAP.h"
 #include "knowhere/index/vector_index/helpers/FaissIO.h"
