+++ core/src/index/knowhere/knowhere/index/vector_offset_index/IndexNSG_NM.cpp
@@ -14,9 +14,9 @@
 
 #include "knowhere/common/Exception.h"
 #include "knowhere/common/Timer.h"
+#include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/IndexIDMAP.h"
 #include "knowhere/index/vector_index/IndexIVF.h"
-#include "knowhere/index/vector_index/IndexType.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
 #include "knowhere/index/vector_index/impl/nsg/NSGIO.h"
 #include "knowhere/index/vector_offset_index/IndexNSG_NM.h"
